backend/requirements.txt
# ═══════════════════════════════════════════════════════════════
# IKSHAN BACKEND — Python Dependencies
# ═══════════════════════════════════════════════════════════════

# Core Framework
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.9.1

# AI & LLM
openai==1.82.0

# HTTP Client (async)
httpx==0.28.1

# Environment
python-dotenv==1.1.0

# Security & Middleware
slowapi==0.1.9
python-multipart==0.0.20

# Logging
structlog==25.1.0

# Performance
orjson==3.10.18

# Document parsing
python-docx==1.1.2

# RAG Pipeline — Vector search for tool recommendations
qdrant-client==1.13.3
numpy==2.2.6

# HTML parsing for website crawling
beautifulsoup4==4.13.3

# ── Phase 2: Research Agent ────────────────────────────────────────────────────
anthropic>=0.30.0

# Redis-backed task streaming (durable events + realtime fanout)
redis>=5.0.0

# Browser automation (used by scraper + backend image installs chromium)
playwright==1.52.0
google-genai>=1.0.0
asyncpg>=0.30.0
playwright
firebase-admin>=6.0.0
PyJWT>=2.8.0
google-auth>=2.0.0
